+++ reactos/lib/msvcrt/stdio/vfwprint.c
@@ -1,5 +1,9 @@
 /* Copyright (C) 1994 DJ Delorie, see COPYING.DJ for details */
-#include <stdarg.h>
+
+//#include <stdarg.h>
+#include <msvcrt/stdarg.h> // robd
+#include <msvcrt/crttypes.h> // robd
+
 #include <msvcrt/stdio.h>
 #include <msvcrt/malloc.h>
 #include <msvcrt/internal/file.h>
@@ -88,15 +92,15 @@ static int skip_wtoi(const wchar_t **s)
 }
 
 
-static int do_div(long long *n,int base)
+static int do_div(LONGLONG *n,int base)
 {
-	int __res = ((unsigned long long) *n) % (unsigned) base;
-	*n = ((unsigned long long) *n) / (unsigned) base;
+	int __res = ((ULONGLONG) *n) % (unsigned) base;
+	*n = ((ULONGLONG) *n) / (unsigned) base;
 	return __res;
 }
 
 
-static int number(FILE * f, long long num, int base, int size, int precision ,int type)
+static int number(FILE * f, LONGLONG num, int base, int size, int precision ,int type)
 {
 	wchar_t c,sign,tmp[66];
 	const wchar_t *digits=L"0123456789abcdefghijklmnopqrstuvwxyz";
@@ -672,7 +676,7 @@ static int stringw(FILE *f, const wchar_t* sw, int len, int field_width, int pre
 int __vfwprintf(FILE *f, const wchar_t *fmt, va_list args)
 {
 	int len;
-	unsigned long long num;
+	ULONGLONG num;
 	int i, base;
 	long double _ldouble;
 	double _double;
@@ -1032,7 +1036,7 @@ int __vfwprintf(FILE *f, const wchar_t *fmt, va_list args)
 		}
 
 		if (qualifier == L'I')
-			num = va_arg(args, unsigned long long);
+			num = va_arg(args, ULONGLONG);
 		else if (qualifier == L'l')
 			num = va_arg(args, unsigned long);
 		else if (qualifier == L'h') {
